requirements.txt
fastapi
uvicorn
chromadb
psycopg2
sentence-transformers
pillow
requests
python-multipart